registerTopic and more Tests
Added the regeisterTopic method and implemented and tested it
primitivley. Because Topics are not to be copied i have decied to change
the signature to take a vector of strings as arguments, instead of
precreated Topics. These Strings are the initalizer variables for the
Topics.

Also the Tests for moveClienttoTopic is now working