Implemented ReadRule
ConfigReader can now Read Rules from the yaml file and handle them
correctly. Tested the Function. If the targetTopic doesn't exist the
rule will be skipped. Think about if this behaviour should be advanced
and used in more situations. You should implement a logger that keeps
track of these Warnings.