Begun Input Deklarations
Begun to define the way Input is handled. After some back and forth the
following design is used: Executable functions and their aruments will
be stored in a Action class which is a wrapper around std::function.
These Objects are then stored in a map, which has a Key or a Button
respectiveley as the Key. These Maps reside in a Singleton Class
InputManager.