Refactor of Action
Action got refactored again. This time for the last one, hopefully.
After studying Desing Patterns for a bit i have decided that a Command
Pattern will be the best. Actions will now be managed with unique ptr,
but much of the old structure could be reused. The handling is now very
similar to the one of Arrangers, but instead of a Selector Class i  am
using a function. I have to redo this a little bit, but the goal is it
to reuse the same vector a bunch of times. THis again screams for a
Singleton.