Implemented DefaultRenderer and drawn first windows
Wind2 is now able to render programs. Currently only in the Monocle
layout and without the closing ability. The Stackorder is respected when
restacking and Raising. Currently no support for floating windows above
all else. The next stepp will be to implement closing windows and then
comes a stability test. Maybe even a first nix derivative.