Implemented Focusstack
I have realized that i need a dedicatet Focusstack. It is now
implemented to stop fullscreen to return the fullscreened window as the
master window. The FUllscreen bug is that if when i change the layout
between spawning and the window actually appearing (xterm is slow) then
the second window can't be focused.